refactor(auth): drop duplicated session-user store + unify profile upsert
- Remove supabaseUserStore / setSupabaseUser: the store mirror of the session
  user was write-only (every reader uses useSupabaseUser() directly), a dead
  second source of truth. useAuth no longer exposes supabaseUser.
- Extract the users-profile upsert into a shared helper
  (composables/Supabase/helpers/user/upsertUserProfile.ts). useAuth and the OAuth
  callback now both call it, removing the triplicated/divergent implementation.
  Unified behaviour: on an existing profile only empty fields are backfilled and
  the role is preserved (the callback previously overwrote name/photo on every
  login). The callback keeps its store write for the non-popup fallback path.

Stages 4 and 8 (slice) of the auth SST plan. Manual OAuth-flow verification
recommended before deploy.
