+++ app/composables/useAuth.ts
@@ -1,5 +1,6 @@
 import { storeToRefs } from 'pinia'
-import type { SupabaseAuthUser, UserInsertData, UserUpdateData } from '~/types/auth'
+import type { SupabaseAuthUser } from '~/types/auth'
+import { upsertUserProfile } from './Supabase/helpers/user/upsertUserProfile'
 
 // Keep auth initialization single-flight across every composable instance.
 let authInitialized = false
@@ -13,22 +14,10 @@ export const useAuth = () => {
 	const { runMutation } = useMutationTimeout()
 
 	// Use storeToRefs to preserve reactivity
-	const { userDataStore, isLoginStore, isAdminStore, supabaseUserStore } =
-		storeToRefs(userStore)
+	const { userDataStore, isLoginStore, isAdminStore } = storeToRefs(userStore)
 
 	// Destructure actions directly; storeToRefs is only needed for refs.
 	const { syncUserProfile, resetStore } = userStore
-
-	const getErrorCode = (error: unknown): string | undefined => {
-		if (typeof error === 'object' && error !== null && 'code' in error) {
-			return (error as { code?: string }).code
-		}
-		return undefined
-	}
-
-	const hasMeaningfulText = (value: string | null | undefined): value is string => {
-		return typeof value === 'string' && value.trim().length > 0
-	}
 
 	const getSessionAuthUser = async (): Promise<SupabaseAuthUser | null> => {
 		const { data } = await supabase.auth.getSession()
@@ -82,143 +71,10 @@ export const useAuth = () => {
 		return await sharedTrustedAuthUserPromise
 	}
 
-	// Create or update a user profile.
-	const createOrUpdateUser = async (authUser: SupabaseAuthUser): Promise<User | null> => {
-		// Supabase v2 can expose a user without an id during initialization.
-		if (!authUser?.id) return null
-
-		try {
-			// Check whether the user already exists.
-			let existingUser: User | null = null
-			let fetchError: { code?: string } | Error | null = null
-
-			if (import.meta.dev) {
-				// Development-only timeout
-				const timeoutPromise = new Promise<never>((_, reject) => {
-					setTimeout(() => reject(new Error('Dev database timeout')), 2000)
-				})
-
-				const fetchPromise = supabase
-					.from('users')
-					.select('*')
-					.eq('id', authUser.id)
-					.single()
-
-				try {
-					const result = await Promise.race([fetchPromise, timeoutPromise])
-					existingUser = result.data as User
-					fetchError = result.error
-				} catch (error) {
-					fetchError = error instanceof Error ? error : new Error('Unknown error')
-				}
-			} else {
-				// Do not use a timeout in production.
-				const result = await supabase
-					.from('users')
-					.select('*')
-					.eq('id', authUser.id)
-					.single()
-
-				existingUser = result.data as User
-				fetchError = result.error
-			}
-
-			if (fetchError && getErrorCode(fetchError) !== 'PGRST116') {
-				console.error("Erreur lors de la récupération de l'utilisateur:", fetchError)
-				throw fetchError
-			}
-
-			if (!existingUser) {
-				// The local user profile creation is also a mutation flow:
-				// if Supabase does not answer, auth must fail clearly instead of hanging.
-				const insertData: UserInsertData = {
-					id: authUser.id,
-					email: authUser.email || '',
-					name:
-						authUser.user_metadata?.full_name ||
-						authUser.user_metadata?.name ||
-						'Utilisateur',
-					photo_url:
-						authUser.user_metadata?.avatar_url || authUser.user_metadata?.picture || '',
-					role: 'USER',
-					created_at: new Date().toISOString(),
-					updated_at: new Date().toISOString(),
-				}
-
-				const { data: newUser, error: createError } = await runMutation(
-					supabase.from('users').insert([insertData]).select().single(),
-					'Creating the user profile timed out. Please try again.',
-				)
-
-				if (createError) {
-					console.error("Erreur lors de la création de l'utilisateur:", createError)
-					const details = JSON.stringify({
-						message: (createError as { message?: string }).message,
-						code: (createError as { code?: string }).code,
-						details: (createError as { details?: string }).details,
-						hint: (createError as { hint?: string }).hint,
-					})
-					throw new Error(`create-user-failed: ${details}`)
-				}
-
-				return newUser as User
-			} else {
-				// Same idea for profile hydration updates: timeout -> explicit error,
-				// no endless "logged in but not really ready" state.
-				const nextEmail = hasMeaningfulText(authUser.email) ? authUser.email : null
-				const nextName =
-					authUser.user_metadata?.full_name || authUser.user_metadata?.name || null
-				const nextPhoto =
-					authUser.user_metadata?.avatar_url || authUser.user_metadata?.picture || null
-
-				const updateData: Partial<UserUpdateData> = {}
-
-				if (!hasMeaningfulText(existingUser.email) && nextEmail) {
-					updateData.email = nextEmail
-				}
-				if (!hasMeaningfulText(existingUser.name) && hasMeaningfulText(nextName)) {
-					updateData.name = nextName
-				}
-				if (!hasMeaningfulText(existingUser.photo_url) && hasMeaningfulText(nextPhoto)) {
-					updateData.photo_url = nextPhoto
-				}
-
-				if (!Object.keys(updateData).length) {
-					return existingUser as User
-				}
-
-				updateData.id = authUser.id
-				updateData.role = existingUser.role
-				updateData.updated_at = new Date().toISOString()
-
-				const { data: updatedUser, error: updateError } = await runMutation(
-					supabase
-						.from('users')
-						.update(updateData)
-						.eq('id', authUser.id)
-						.select()
-						.single(),
-					'Updating the user profile timed out. Please try again.',
-				)
-
-				if (updateError) {
-					console.error("Erreur lors de la mise à jour de l'utilisateur:", updateError)
-					const details = JSON.stringify({
-						message: (updateError as { message?: string }).message,
-						code: (updateError as { code?: string }).code,
-						details: (updateError as { details?: string }).details,
-						hint: (updateError as { hint?: string }).hint,
-					})
-					throw new Error(`update-user-failed: ${details}`)
-				}
-
-				return updatedUser as User
-			}
-		} catch (error) {
-			console.error('Erreur dans createOrUpdateUser:', error)
-			throw error
-		}
-	}
+	// Create or update the application profile. Delegates to the shared helper so
+	// this composable and the OAuth callback never diverge in their upsert logic.
+	const createOrUpdateUser = (authUser: SupabaseAuthUser): Promise<User | null> =>
+		upsertUserProfile(supabase, authUser, runMutation)
 
 	// Synchronization state
 	const isSyncing = useState('auth-is-syncing', () => false)
@@ -226,8 +82,8 @@ export const useAuth = () => {
 
 	const preserveAuthenticatedState = (authUser: SupabaseAuthUser) => {
 		// Mark the client as authenticated immediately while profile hydration
-		// catches up in the background.
-		userStore.setSupabaseUser(authUser)
+		// catches up in the background. The live session user comes from
+		// useSupabaseUser(); we no longer mirror it into the store.
 		userStore.setIsLogin(true)
 
 		// isAdmin derives from the profile automatically; only drop a stale profile
@@ -431,7 +287,6 @@ export const useAuth = () => {
 		isAdmin: isAdminStore,
 		isLoggedIn,
 		isReady,
-		supabaseUser: supabaseUserStore,
 		isSyncing: readonly(isSyncing),
 		syncError: readonly(syncError),
